refactor: split monolithic test file into focused test modules + add requirement page type tests
- Split test/llm-wiki.test.ts (950+ lines) into 6 focused test files:
  - test/package-structure.test.ts — package.json, SKILL.md, prompts, tools
  - test/source-capture.test.ts — URL, PDF, file, XML, JSON capture
  - test/wiki-structure.test.ts — vault dirs, page types (incl. requirements)
  - test/xref-config.test.ts — cross-references, config, backward compat
  - test/recall.test.ts — wiki_recall search and formatting
  - test/retro.test.ts — wiki_retro insight saving
- Created test/helpers.ts with shared utilities (no module-level state)
- Added requirement page type tests: frontmatter, dependency tracking,
  status lifecycle, metadata discovery, wiki_recall discoverability
- Coverage: 61 tests, same logical coverage as original 55 + 6 new
